visual-injection-prompt/requirements.txt
openai
scikit-learn
joblib
python-dotenv
